requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
fastapi==0.92.0
Flask==2.2.3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pydantic==1.10.5
pymongo==4.3.3
python-multipart==0.0.6
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.3
yapf==0.32.0
